+++ src/client/renderer-new/Boilerplate.cpp
@@ -4,6 +4,8 @@
 #include "Boilerplate.hpp"
 
 #include <filesystem>
+
+#define STB_IMAGE_IMPLEMENTATION
 #include <stb_image.h>
 
 namespace Boilerplate
